run each tests case in isolation
