+++ test/jasmine/performance_tests/all_test.js
@@ -3,6 +3,7 @@ var delay = require('../assets/delay');
 var d3SelectAll = require('../../strict-d3').selectAll;
 var Plotly = require('../../../lib/index');
 var downloadCSV = require('./assets/post_process').downloadCSV;
+var tests = require('./assets/test_cases').testCases;
 var nSamples = require('./assets/constants').nSamples;
 var MAX_RENDERING_TIME = 4000;
 
@@ -10,78 +11,6 @@ var gd = createGraphDiv();
 
 const samples = Array.from({ length: nSamples }, (_, i) => i);
 
-var tests = [];
-
-for(let traceType of ['image', 'heatmap', 'contour']) {
-    for(let m of [10, 20, 40, 80, 160, 320, 640]) {
-        let nx = 5 * m;
-        let ny = 2 * m;
-        tests.push({
-            nx: nx,
-            ny: ny,
-            n: nx * ny,
-            nTraces: 1,
-            traceType: traceType,
-            selector: traceType === 'image' ? 'g.imagelayer.mlayer' :
-                'g.' + traceType + 'layer'
-        });
-    }
-}
-
-for(let traceType of ['box', 'violin']) {
-    for(let mode of ['no points', 'all points']) {
-        for(let nTraces of [1, 10, 100]) {
-            for(let n of [1000, 2000, 4000, 8000, 16000, 32000]) {
-                tests.push({
-                    n:n,
-                    nTraces: nTraces,
-                    traceType: traceType,
-                    mode: mode,
-                    selector: (
-                        traceType === 'box' ? 'g.trace.boxes' :
-                        traceType === 'violin' ? 'g.trace.violins' :
-                        undefined
-                    )
-                });
-            }
-        }
-    }
-}
-
-for(let traceType of ['scatter', 'scattergl', 'scattergeo']) {
-    for(let mode of ['markers', 'lines', 'markers+lines']) {
-        for(let nTraces of [1, 10, 100]) {
-            for(let n of [1000, 2000, 4000, 8000, 16000, 32000]) {
-                tests.push({
-                    n:n,
-                    nTraces: nTraces,
-                    traceType: traceType,
-                    mode: mode,
-                    selector: (
-                        traceType === 'scatter' ? 'g.trace.scatter' :
-                        undefined
-                    )
-                });
-            }
-        }
-    }
-}
-
-for(let traceType of ['bar', 'histogram']) {
-    for(let mode of ['group', 'stack', 'overlay']) {
-        for(let nTraces of [1, 10, 100]) {
-            for(let n of [1000, 2000, 4000, 8000, 16000, 32000]) {
-                tests.push({
-                    n:n,
-                    nTraces: nTraces,
-                    traceType: traceType,
-                    mode: mode,
-                    selector: 'g.trace.bars'
-                });
-            }
-        }
-    }
-}
 
 function generateMock(spec) {
     var type = spec.traceType;
@@ -161,17 +90,17 @@ function makeBox(spec) {
     for(var k = 0; k < spec.nTraces; k++) {
         var trace = {
             type: spec.traceType,
-            boxpoints: spec.mode === 'all points' ? 'all' : false,
+            boxpoints: spec.mode === 'all_points' ? 'all' : false,
             y: y.slice(k * nPerTrace, (k + 1) * nPerTrace),
             x: Array.from({ length: nPerTrace }, (_, i) => k)
         };
 
         if(spec.traceType === 'box') {
-            trace.boxpoints = spec.mode === 'all points' ? 'all' : false;
+            trace.boxpoints = spec.mode === 'all_points' ? 'all' : false;
         }
 
         if(spec.traceType === 'violin') {
-            trace.points = spec.mode === 'all points' ? 'all' : false;
+            trace.points = spec.mode === 'all_points' ? 'all' : false;
         }
 
         data.push(trace);
@@ -283,8 +212,8 @@ describe('Performance test various traces', function() {
         if(testCase) {
             if(testCase.tracesType && testCase.tracesType !== spec.traceType) testIt = false;
             if(testCase.tracesCount && testCase.tracesCount !== spec.nTraces) testIt = false;
-            if(testCase.tracesMode && testCase.tracesMode !== spec.mode) testIt = false;
             if(testCase.tracesPoints && testCase.tracesPoints !== spec.n) testIt = false;
+            if((testCase.tracesMode !== 'undefined' && testCase.tracesMode) && testCase.tracesMode !== spec.mode) testIt = false;
         }
 
         if(testIt) {
